update auto subsystem to put variables before __init__
