+++ src/subsystems/autoRoutines.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class AutoRoutines(Enum):
+    DO_NOTHING = "Do Nothing"
+    DRIVE_FORWARD_TEST = "Drive Forward Test"
+    DRIVE_FORWARD_BACK_TEST = "Drive Forward Back Test"
+    WONKY = "Wonky"
